refactor: Extract PasswordInput and validation to stay under 300-line limit
Move PasswordInput component (with show/hide toggle, strength bar,
hint text) to registration-helpers.tsx and field validation logic to
validateRegistrationFields() in registration-utils.ts. Keeps
register-page.tsx at 295 lines (limit: 300).
