+++ frontend/src/features/registration/pages/register-page.tsx
@@ -1,14 +1,13 @@
 import { useState, useCallback, useEffect, useRef, type FormEvent } from 'react'
 import { useNavigate, Link } from 'react-router-dom'
-import { validateSlug, isSafeRedirectUrl, type SlugAvailability } from './registration-utils'
-import { PasswordStrengthBar, SlugStatus, RedirectSuccess } from './registration-helpers'
+import { validateSlug, validateRegistrationFields, isSafeRedirectUrl, type SlugAvailability } from './registration-utils'
+import { PasswordInput, SlugStatus, RedirectSuccess } from './registration-helpers'
 
 export function RegisterPage() {
   const navigate = useNavigate()
   const [slug, setSlug] = useState('')
   const [email, setEmail] = useState('')
   const [password, setPassword] = useState('')
-  const [showPassword, setShowPassword] = useState(false)
   const [displayName, setDisplayName] = useState('')
   const [slugError, setSlugError] = useState<string | null>(null)
   const [slugAvailability, setSlugAvailability] = useState<SlugAvailability>('idle')
@@ -49,7 +48,6 @@ export function RegisterPage() {
     setSlugAvailability('checking')
 
     const timer = setTimeout(() => {
-      // Abort any in-flight request
       slugCheckController.current?.abort()
       const controller = new AbortController()
       slugCheckController.current = controller
@@ -80,41 +78,17 @@ export function RegisterPage() {
     }
   }, [slug])
 
-  const handleSlugChange = useCallback((value: string) => {
-    setSlug(value.toLowerCase().replace(/[^a-z0-9-]/g, ''))
-    setFieldErrors((prev) => {
-      const next = { ...prev }
-      delete next.slug
-      return next
-    })
+  const clearFieldError = useCallback((field: string) => {
+    setFieldErrors((prev) => { const next = { ...prev }; delete next[field]; return next })
   }, [])
 
+  const handleSlugChange = useCallback((value: string) => {
+    setSlug(value.toLowerCase().replace(/[^a-z0-9-]/g, ''))
+    clearFieldError('slug')
+  }, [clearFieldError])
+
   const validateFields = useCallback((): boolean => {
-    const errors: Record<string, string> = {}
-
-    if (!slug) {
-      errors.slug = 'Organization slug is required'
-    } else {
-      const slugValidation = validateSlug(slug)
-      if (slugValidation) {
-        errors.slug = slugValidation
-      }
-    }
-
-    const normalizedEmail = email.trim().toLowerCase()
-    const emailPattern = /^[^\s@]+@[^\s@]+\.[^\s@]+$/
-    if (!normalizedEmail) {
-      errors.email = 'Email is required'
-    } else if (!emailPattern.test(normalizedEmail)) {
-      errors.email = 'Please enter a valid email address'
-    }
-
-    if (!password) {
-      errors.password = 'Password is required'
-    } else if (password.length < 8) {
-      errors.password = 'Password must be at least 8 characters'
-    }
-
+    const errors = validateRegistrationFields(slug, email, password)
     setFieldErrors(errors)
     return Object.keys(errors).length === 0
   }, [slug, email, password])
@@ -171,14 +145,11 @@ export function RegisterPage() {
         const loginUrl = typeof data?.login_url === 'string' ? data.login_url : undefined
 
         if (loginUrl && isSafeRedirectUrl(loginUrl)) {
-          // Safe tenant subdomain URL - show success then redirect
           setRedirecting(true)
           redirectTimerRef.current = window.setTimeout(() => {
             window.location.href = loginUrl
           }, 1500)
         } else {
-          // Relative path, missing, or untrusted URL - use client-side navigation
-          // Reject absolute URLs that failed validation to avoid broken SPA routes
           const fallbackPath = (loginUrl && !loginUrl.startsWith('http')) ? loginUrl : '/login?registered=1'
           void navigate(fallbackPath)
         }
@@ -273,11 +244,7 @@ export function RegisterPage() {
               value={email}
               onChange={(e) => {
                 setEmail(e.target.value)
-                setFieldErrors((prev) => {
-                  const next = { ...prev }
-                  delete next.email
-                  return next
-                })
+                clearFieldError('email')
               }}
               required
               autoComplete="email"
@@ -290,61 +257,16 @@ export function RegisterPage() {
             )}
           </div>
 
-          <div>
-            <label htmlFor="password" className="block text-sm font-medium mb-1">
-              Password <span className="text-destructive" aria-hidden="true">*</span>
-              <span className="sr-only">(required)</span>
-            </label>
-            <div className="relative">
-              <input
-                id="password"
-                type={showPassword ? 'text' : 'password'}
-                value={password}
-                onChange={(e) => {
-                  setPassword(e.target.value)
-                  setFieldErrors((prev) => {
-                    const next = { ...prev }
-                    delete next.password
-                    return next
-                  })
-                }}
-                required
-                autoComplete="new-password"
-                aria-describedby="password-strength password-hint"
-                aria-invalid={submitted && !!fieldErrors.password ? true : undefined}
-                className={`${submitted && fieldErrors.password ? inputErrorClass : inputClass} pr-10`}
-                minLength={8}
-              />
-              <button
-                type="button"
-                onClick={() => setShowPassword((v) => !v)}
-                className="absolute inset-y-0 right-0 flex items-center pr-3 text-muted-foreground hover:text-foreground"
-                aria-label={showPassword ? 'Hide password' : 'Show password'}
-              >
-                {showPassword ? (
-                  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" className="h-4 w-4" aria-hidden="true">
-                    <path fillRule="evenodd" d="M3.28 2.22a.75.75 0 00-1.06 1.06l14.5 14.5a.75.75 0 101.06-1.06l-1.745-1.745a10.029 10.029 0 003.3-4.38 1.651 1.651 0 000-1.185A10.004 10.004 0 009.999 3a9.956 9.956 0 00-4.744 1.194L3.28 2.22zM7.752 6.69l1.092 1.092a2.5 2.5 0 013.374 3.373l1.092 1.092a4 4 0 00-5.558-5.558z" clipRule="evenodd" />
-                    <path d="M10.748 13.93l2.523 2.523a9.987 9.987 0 01-3.27.547c-4.258 0-7.894-2.66-9.337-6.41a1.651 1.651 0 010-1.186A10.007 10.007 0 012.839 6.02L6.07 9.252a4 4 0 004.678 4.678z" />
-                  </svg>
-                ) : (
-                  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" className="h-4 w-4" aria-hidden="true">
-                    <path d="M10 12.5a2.5 2.5 0 100-5 2.5 2.5 0 000 5z" />
-                    <path fillRule="evenodd" d="M.664 10.59a1.651 1.651 0 010-1.186A10.004 10.004 0 0110 3c4.257 0 7.893 2.66 9.336 6.41.147.381.146.804 0 1.186A10.004 10.004 0 0110 17c-4.257 0-7.893-2.66-9.336-6.41zM14 10a4 4 0 11-8 0 4 4 0 018 0z" clipRule="evenodd" />
-                  </svg>
-                )}
-              </button>
-            </div>
-            <p id="password-hint" className="mt-1 text-xs text-muted-foreground">
-              Minimum 8 characters.
-            </p>
-            {submitted && fieldErrors.password ? (
-              <p className="mt-1 text-xs text-destructive" role="alert">{fieldErrors.password}</p>
-            ) : (
-              <div id="password-strength" aria-live="polite">
-                <PasswordStrengthBar password={password} />
-              </div>
-            )}
-          </div>
+          <PasswordInput
+            value={password}
+            onChange={(e) => {
+              setPassword(e.target.value)
+              clearFieldError('password')
+            }}
+            inputClass={inputClass}
+            error={fieldErrors.password}
+            submitted={submitted}
+          />
 
           {formError && (
             <p role="alert" className="text-sm text-destructive">
