+++ frontend/src/features/registration/pages/registration-helpers.tsx
@@ -1,3 +1,4 @@
+import { useState, type ChangeEvent } from 'react'
 import { passwordStrength, type SlugAvailability } from './registration-utils'
 
 const STRENGTH_LABEL = ['', 'Weak', 'Fair', 'Good', 'Strong']
@@ -62,4 +63,77 @@ export function SlugStatus({ slug, error, availability }: SlugStatusProps) {
     default:
       return <p className="mt-1 text-xs text-green-600">Slug format looks good</p>
   }
+}
+
+function EyeIcon() {
+  return (
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" className="h-4 w-4" aria-hidden="true">
+      <path d="M10 12.5a2.5 2.5 0 100-5 2.5 2.5 0 000 5z" />
+      <path fillRule="evenodd" d="M.664 10.59a1.651 1.651 0 010-1.186A10.004 10.004 0 0110 3c4.257 0 7.893 2.66 9.336 6.41.147.381.146.804 0 1.186A10.004 10.004 0 0110 17c-4.257 0-7.893-2.66-9.336-6.41zM14 10a4 4 0 11-8 0 4 4 0 018 0z" clipRule="evenodd" />
+    </svg>
+  )
+}
+
+function EyeSlashIcon() {
+  return (
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" className="h-4 w-4" aria-hidden="true">
+      <path fillRule="evenodd" d="M3.28 2.22a.75.75 0 00-1.06 1.06l14.5 14.5a.75.75 0 101.06-1.06l-1.745-1.745a10.029 10.029 0 003.3-4.38 1.651 1.651 0 000-1.185A10.004 10.004 0 009.999 3a9.956 9.956 0 00-4.744 1.194L3.28 2.22zM7.752 6.69l1.092 1.092a2.5 2.5 0 013.374 3.373l1.092 1.092a4 4 0 00-5.558-5.558z" clipRule="evenodd" />
+      <path d="M10.748 13.93l2.523 2.523a9.987 9.987 0 01-3.27.547c-4.258 0-7.894-2.66-9.337-6.41a1.651 1.651 0 010-1.186A10.007 10.007 0 012.839 6.02L6.07 9.252a4 4 0 004.678 4.678z" />
+    </svg>
+  )
+}
+
+interface PasswordInputProps {
+  value: string
+  onChange: (e: ChangeEvent<HTMLInputElement>) => void
+  inputClass: string
+  error?: string | null
+  submitted?: boolean
+}
+
+export function PasswordInput({ value, onChange, inputClass, error, submitted }: PasswordInputProps) {
+  const [showPassword, setShowPassword] = useState(false)
+  const hasError = submitted && !!error
+  const errorClass = hasError ? inputClass + ' border-destructive' : inputClass
+
+  return (
+    <div>
+      <label htmlFor="password" className="block text-sm font-medium mb-1">
+        Password <span className="text-destructive" aria-hidden="true">*</span>
+        <span className="sr-only">(required)</span>
+      </label>
+      <div className="relative">
+        <input
+          id="password"
+          type={showPassword ? 'text' : 'password'}
+          value={value}
+          onChange={onChange}
+          required
+          autoComplete="new-password"
+          aria-describedby="password-strength password-hint"
+          aria-invalid={hasError ? true : undefined}
+          className={`${errorClass} pr-10`}
+          minLength={8}
+        />
+        <button
+          type="button"
+          onClick={() => setShowPassword((v) => !v)}
+          className="absolute inset-y-0 right-0 flex items-center pr-3 text-muted-foreground hover:text-foreground"
+          aria-label={showPassword ? 'Hide password' : 'Show password'}
+        >
+          {showPassword ? <EyeSlashIcon /> : <EyeIcon />}
+        </button>
+      </div>
+      <p id="password-hint" className="mt-1 text-xs text-muted-foreground">
+        Minimum 8 characters.
+      </p>
+      {hasError ? (
+        <p className="mt-1 text-xs text-destructive" role="alert">{error}</p>
+      ) : (
+        <div id="password-strength" aria-live="polite">
+          <PasswordStrengthBar password={value} />
+        </div>
+      )}
+    </div>
+  )
 }
